safari-600.7-branch/LayoutTests/platform/mac/editing/input/maxlength-expected.txt

Test for bug 19479: Special characters can be inserted in text field having reached maxlength.


PASS testElement.value is "12345678"
PASS testElement.value is "12345678"
PASS successfullyParsed is true

TEST COMPLETE
